Add files via upload
In this project I've learn a lot of things. I've learn that in general  for style the page it is  important to divide the parts, with div, is it important to give class, and then use all the important tag to underline the header ,section, fieldset, footer ecc. I've learn that in CSS is it important to use the correct format, for exaple PX for what will not change in the page, or % for elements that are inside a container and so will change. I'm still learning but I think this was a usfuel project.